working on refactoring of the CallsQueueOperator source code. Refactoring will allows to make parallel calls to the operator's numbers

+++ trunk/onesec-raven/src/main/java/org/onesec/raven/ivr/queue/CallsCommutationManager.java
@@ -29,7 +29,7 @@ public interface CallsCommutationManager {
     public void callFinished(CommutationManagerCall call, boolean successfull);
     public CallsQueueOperator getOperator();
     public long getWaitTimeout();
-    public int getInviteTimeout();
+    public long getInviteTimeout();
     public IvrConversationsBridgeManager getConversationsBridgeManager();
     public IvrEndpointPool getEndpointPool();
     public IvrConversationScenario getConversationScenario();
